refactor(charm): drop dangerous ReadCharm and ReadBundle
These generic functions look at a path, determin if the resource is an
archive or dir, and then run the associated parser.

However, ever since we dropped support for deploying charms from
directories, we have always known ahead of time the format of what we
are about to read.

In fact, we never need to parse a full charm as a directory. At most, we
need to read one particular charm file. The only exception are in tests.
So move ReadCharmDir into a testing package.

The rest of the PR is made up of refactors/simplifications which
were made possible by this change.