feat: add requires sequencing flag
By indicating that the charm requires sequencing we make it very
explicit that the charm needs to be given a new revision from the
sequence table. It is then upto the service to ensure correctness.
Sequencing isn't some implicit set of values; local charm source
with the revision set to -1. Just call it out for what it is.