ci: fix rtd build
The sphinx documentation is automatically built by RTD using the
template in the .readthedocs.yaml file. Because we auto generate the CLI
documentation using the juju documentation command, it the juju command
itself to be built. To do this requires some hackery.

The RTD build process is overwritten using the commands config option.
In it, we build juju, then make the docs with the juju executable on the
path.