chore: min charm sha256 prefix constant
Unfortunately, we've made a boo boo. Essentially, the api client
wants the sha256 to be 8 characters long, and the api server wants
it to be 7. We I've said now is that we require a minimum of 7 and
you can provide the full hash if you like. This will then allow us
to move towards a better s3 object compatibility. So win win in the
end!