fix: returned uuid when inserting charm with same hash
This fixes an issue presented when we tried to deploy the same charm
twice (or the same charm from charmhub and then from local). The error
was that in the case of a conflict in the hash, the same line in the db
was used for the object store metadata, but the returned UUID was still
a new one. Now the UUID corresponding to that line gets returned.