fix: migration failed when charms where uploaded
This commit fixes two different issues that were preventing migrations
to work:
- The hash prefix length that was passed on when PUTting the charm was 7
  chars length, and not 8. See internal/migration/migration.go
- The application service getter was using the incorrect service during
  migration. This happened because since it peeked at the model uuid
from the request context, then the controller UUID was returned and
therefore the charms weren't found. The fix was to add a different
service getter for migrating operations, exactly as it was done with the
legacy state during migration.