fix: avoid checking wrong URL assignment for local charms migration
We have added the architecture as part of the charm URL returned when
uploading a local charm, which makes the migration fail because the
serialized application in the model contains the URL without it.
This commit therefore removes this check only for local charms, with a
note that it can be reinstated when we serialize the full application
from the domain.