chore: cleanup legacy state charm methods
This patch removes most of the legacy Charm methods, as well as private
methods and structs used by these methods.
This can be done without major issues because we were not double writing
anymore at this point.

The only public method that is left for future cleanup is
`LatestPlaceholderCharm()` which is still used in the status (although
no data is returned), so it can be reimplemented correctly using the
charm domain in a future patch.