feat: only include arch when requested
During migration we only want to include the architecture when the
client has sent it. This allows us to migrate away from
architecture-less URLs. Thus a 4.x can always send the architecture
and the 3.x client can send it optionally. It's a bit of a hack, but
it prevents us from forcing a new release of a 3.x client.

If the architecture is missing in the URL, then we'll use the current
strategy, so it's not a problem, it's just a lack of data.