feat: get applications for revision updater
The revision updater needs specific information so it can update
the revisions of a charm. We use views to reduce the complexity of
the queries to get the information. We can't merge the two queries
for the revision updater application and the number of application
units, as the units can be optional (you can have an application
without units).

Also as a driveby I fixed the database testing method, it was missing
a new line to correctly output the information.