chore: remove charm payloads
This removes charm payloads from juju. It's not currently utilised
and it will require migrating from mongo to dqlite. The concept
of a payload was to model on machines (IAAS only) that an assoicated
service was hosted on another service (docker for example). Something
like this might be useful, but it wouldn't be implemented in the
same way, so rather than migrate over dead code, we should create
something that meets the needs of the user. In any event, we could
always look at this code in the git history.

The APIs for v1 data will just return empty or canned data and the
v2 of the APIs are blackholed. We can't remove them completely, as
the agent won't come up after a 3.6 migration.