feat: wire up resource opener to resource service
Wire up the resource opener to the resource service. The resource opener
is used by the http resource api and the caas application provisioner
worker.

Changes:
 - In 3.6, when a resource had finished being read, it would be set as in
   use in state. This was achived by writing to state from inside the
   io.Read method. To replicate this behviour more explicitly in 4.0, a
   method, SetResource has been added to the resource opener, this is used
   by the callers of opener.OpenResource to mark that the resource is in
   use.

Note that the resource opener still uses state to get the charmURL, this
is because it is not yet available from the application domain. A note
has been added to fix this when it becomes available.