test: add tests for NewResourceOpeners
The methods NewResourceOpener and NewResourceOpenerForApplication were
previously untested but contained significant business logic. This
involves the changes below.

The parts of the internal/resource package dealing with charmhub have
been refactored into internal/resource/charmhub, put benhind
interfaces and injected to allow for effective testing. This has
meant the dependency on the ModelConfigService can be removed.

To test the state, a state shim was added that returned the interface
types needed for the shim to fulfil state interface. The original state
is now passed to the `New` methods which then wrap it in the shim and
call private new methods.

Previously the New methods were tested by calling the public New method,
now it uses the private one since the public one cannot accept a mocked
state (the previous implemetation used the mocked state on the public
one, bud did not work when wired up).

To test the private `New` methods, the tests have been put in the
resources package, rather than the resources_test package. This allows
them to access it directly without the need for export_test.go.

A name clash with the global logger variable and a logger declared in
the init of an imported package name meant that the logger had to be
renamed.