chore: remove status history
Status history will not be written to dqlite, instead it will be
off loaded to the logging infrastructure. To prevent the application
and unit dqlite migration adding code that will be deleted, we're
just going to remove the code and implement this in the status
history epic. There is prior work to sound out status history[1]
with the aim to provide the same functionality.

 1. https://github.com/juju/juju/pull/17237