feat: wire up http resource handler
The HTTP resource handler previously used state to serve and save
resources. This change uses pipes in the resource service and uses that
instead.

The response sent back from successful PUT requests has been ignored in
the client since at least Juju 2.x. For this reason, the response is
to successful put requests is now empty.