feat: store new revisions
Once the charm revisioner worker has run, we need to store the new
revisions. These revisions need to contain all the essential
metadata, so when the charm is swapped to the new one, it can just
be downloaded with all the correct data. Effectively the charm
is in all circumstances deployed.

Reworking the charmrevisioner to remove the charm URL and just
use the charm locator removes a needless encoding of the charm
URL.

The architecture is tricky, because it should just be the same as
the existing application, but there is no way to verify that using
the response. We just have to trust the API.