feat: remote api caller
This code is part of the work to enable file based replication. It's
a considerable piece of work that should be reviewed on it's own
merits. Most of the code here is lifted out of the pubsub worker and
generalized for use in other locations. Once the file based
replication work has been done, it is expected that the pubsub worker
will also transition over to this worker.

The concept for the api remote caller is to hold a series of api
connections to other controller nodes. Currently there connection
is not monitored for a broken connection, that can be done later to
improve the resilience of the connections.

This doesn't establish a connection to ourselves, only to other
controller nodes. Thus making a mesh network.

The difference between the pubsub worker and this new worker, is
that we're leaning heavily on the runner package to keep the connections
running.