fix: close state pool item on release
This closes the state pool item on the release and removal of the
pooled state. There is a logical race between when a model is
removed and when another item requests the same model. Unfortunately,
what happens then, is that nothing calls remove on the pool item,
only an item is released. The side-effect of this is that pool items
are kept perpetuity. This can be witnessed when adding a model,
then removing a model and then inspecting the juju_engine_report.
The state field contains models that have been destroyed. The
models have been removed, the death throws of removal of a model
spawns new calls to the state pool, which ultimately go away.

There is a potential issue, that when we call StatePool.Get, we
reopen the state, and it's associated workers (txn log watcher),
thus creating churn, but this is short lived and everything is
correctly closed.

Alternatively, we could stop any StatePool.Get requests after a
StatePool.Remove request is called, using the tombstone map, but
it's possible (although slim) that a removal is possible and access
might still be required. It's therefore "safer" to allow the closing
of a pool on the release if the reference count is zero.

The code employs the tombstone map to ensure that when you release
a pool after aquiring it, that the remove method still holds the
guarantee that it returns true if it's been removed. It doesn't
stop you from re-accessing the pool again.