docs: remove auto-generate CLI docs from tree
This folder has been added to the docs .gitignore in a previous commit,
but the docs in it were not removed from the tree. These docs are
docs are generated fresh every time the docs are built.

Since this folder is ignored by git, these files will never change and
should not be in the tree.