feat: add size and fingerprint parameters to store resource
The size and fingerprint of the resource should be calculated when the
resource is recieved by the controller. These are validated by the
object store when the resource is read in (though this is only for file
resources), it should however be done earlier, before the resource gets
to the serivce layer.

This change records the expected size and fingerprint of the resource in
storage. These values are expected to be validated outside of the the
service.