refactor: rename environschema to configschema
Although environschema was originally intended to be used for
environs, it's managed to spread throughout the codebase of juju and
has become predominantly configuration schema for different things.

The configschema isn't the ideal abstraction, we would ideally
remove it and just used types for model config and application config,
but that will require additional time. Part of that work was started
by adding a schema.List[1], but being a library that others also
depend on means we can't be as ruthless about the changes that Juju
requires. By bringing the dependency into Juju, we can start making
more agressive changes for the needs of Juju, without having to
consider the usecases of other libraries.

 1. https://github.com/juju/environschema/pull/16