refactor: drop environschema as a direct dep
Now that the idmclient offers a form.Filler constructor for creating
a IO filler, we don't need to construct one from environschema
directly. This import gymnastics just removes the need to manage
the environschema directly from juju.

When we remove macaroons from Juju, we can remove this nicely with
a clean cut.