test: fix intermittent failure in TestRecallUnit
The stub relied upon to indicate that a unit is started by the deployer
returns too early. This can result in an error from unit recall due to
files on disk still being modified when we attempt to remove it.

This checks that the introspection socket file is present before
attempting to proceed with recall.