chore: backport pull request #18629 from manadart/fix-recall-unit-test
https://github.com/juju/juju/pull/18629

In the deployer tests, we have `unitWorkersStub`, which is used as the `Manifold` for the unit agent dependency engine.

This effectively represents a unit's workers as started the moment its `Start` function is called. It is *after* this point that the introspection socket is created for the new agent on disk.

This in turn may cause problems in `TestRecallUnit`, because we only check for the started condition above before attempting to recall. `RecallUnit` may then attempt to delete the unit's agent directory while we are creating the socket.

This likely became possible after we switched away from abstract sockets due to security concerns.

Here we simply check for the socket file presence before proceeding.

Run `TestRecallUnit` from /internal/worker/deployer.go in a loop.