feat: get latest pending charmhub charm
The state method returns the latest pending charmhub charm. We
can't use the revision to determin order, as charm revisions can
be out of order; think of multiple channels, that promote themselves
to different channels. A stable channel might have revision 1, then
multiple updates to candidate, ending with revision 20, which is
then promoted to stable. Then revision 1 becomes 20, and it's possible
to promote a prior revision to stable; 19 for instance.