feat: wire up the latest pending charms
Wires up the latest pending charms call to the status. We don't
directly expose the can upgrade to flag, although it is available
when calling a different formatter (yaml or json).

I've also fixed the issue when it was looking for a revision greater
than the current one. This is incorrect, revision can appear out of
order. In any case we just want a version that is greater than the
current one.