feat: use Download instead of DownloadResource in resource charmhub client
Switch to using the Download method of the charmhub client in the
resource charmhub client rather than the DownloadResource method becuase
Download calculates the hash and size of the downloaded blob, these can
then be used in the resource charmhub client to verify the against the
expected size and hash.

The Download function writes to a given file rather than returning an
io.Reader. For this reason we create a temporary file and return a
wrapper around it that also deletes the file when the io.Reader is
closed.