chore: add --force to destroy_model() in bash tests
Currently, destroy-model is broken and it gets stuck without any errors
and without removing the model unless it's used with --force.
This is a big issue on our bash tests because most of them fail because
of cleanup and it hides the actual errors underneath.
This patch therefore adds --force to destroy_model() in our bash tests
and allows us to fix the actual errors. This change is only done at one
place so it should be easy to reinstate after cleanup is fixed.