feat: remove min units and application scaler
MinUnits was originally intended to be set during ApplicationUpdate,
but that facade call was removed from 2.9 to 3.0. Yet we left all
the workers around that responded to that value changing. It's
potentially possible to do this in 3.6 branch and then forward
port it to main, but there are some tricky intricacies around the
all watcher and depreacting fields.

For now, we can rip out all of this in the main branch because
there is no all watcher, neither is there going to be any state
code.