feat: ensure we correctly cancel with new address
This is a bit tricky, as you have unlimited attempts and we want to
cancel them when we get a new change. We don't want to store the
context on the server, but we can pass it through with the change
request. Thus allowing us to do correctly cancel and remove all
goroutines on a defer.