feat: add state methods for application config
Adds the state methods for getting the application config. We don't
have the liberty of doing the request in one query. We need to know
if the application exists, if the trust setting is correctly set, and
then get the application config. It's possilbe to maybe UNION the
application config and settings in one view, but it gets messy if
there are no configs when attempting the UNION. Instead, we can keep
the statements small and simple and we can optimize if we ever need
to.