fix: use fingerprint in path for stored resource
The resource used to be stored in the objectstore/container image store
via its uuid. This posed an issue when updating a resource. If a
resource was already in the store it would need to be removed before the
new one could be added or an error would be thrown.

This change means that the fingerprint is also used in the key, thus
making it unique for a unique resource file.