fix(apt): sources location & format changes in Noble
Starting with noble, the location and format for the default ubuntu apt
sources changed.

This resulted in some of our bash scripts breaking, as they are coded
the apt sources file path, and assumed a particular format.

Re-write these script to work for distros both before and after this
change