feat: allow resource opener to give up
If the resource opener has a limit set, either globally or for an
application, we want to be able to tell the opener to give up. Either
by creating a timeout, or by the client closing. Either way, the
original implementation didn't allow this, it would just block forever.
We can swap out the channels for the weighted semaphore, which will
allow the nice and easy cancellation of resources.