fix: do not return the resource to drop if old is the same as new
StoreResource returns the fingerprint of any old resource that needs to
be dropped. If the resource being recorded was the same as the resource
already there, there is a chance it would return the UUID and this would
get dropped meaning there was no resource blob. Handle this case by
returning no fingerprint to drop in it.