feat: remove leader-settings
The leader-settings have been deprecated for sometime and we're
finally removing them in 4.0. The operator framework has a
deprecated warning on the HookEvent[1] for some time. Only
is_leader is available, which is still accessible after this
code removal.

The code around the uniter is interesting, as we need to keep
the API for v20 (3.6 jujud agent), so after we've migrated it
still comes up without crashing[2] upon starting. For the
v21 version of the API we can just blank out the methods.
The v20 version of the API will just return either no information
or a watcher that won't do anything.

 1. https://github.com/canonical/operator/blob/23eced953fb68ffb9af4889581512fe51e6c61bd/ops/charm.py#L411-L418
 2. https://github.com/juju/juju/blob/11d1f21773373d22eb7dee6cbb93c08027e811c2/apiserver/facades/agent/uniter/register.go#L28-L30