feat(relation): stub the relation domain and add to factory
Stub the relation service and state. Includes the initial service
methods with the expectation that the errors will be updated as they are
implemented. Each service method will return an error of NotImplemented
where an error is returned, -1 for ints and false for booleans.

State methods will be defined as the service methods are implemented.

Unit tests will be added as the domain is implemented.

Add to the model services, so we won't forget later.