feat: set application config
Allow the setting of the application config based on the new values
vs the old values. This updates any values and types correctly,
whilst inserting and removing. This also handles the updating of
the trust setting.

The trust setting is always returned even if there is no row for
it.