feat: application config service methods
Adds the application config service methods. This includes getting
and setting the application config. The config will be in the form
of a map[string]string, with the yaml file from the apiserver being
merged before it hits the service methods. This ensures that there
is one consistent view within the service methods.