feat: allow unsetting of application config keys
We want to be able to unset and remove configuration keys on an
application. We don't want to go through the SetApplicationConfig,
as that requires loading up the current config so that we can remove
keys. This will just remove them.

We don't care if they're config vs settings in this case, as these
are just keys.