feat: default container type for constraint
This commit introduces both a default value for a model's container
straint and a contract assertion around the values for a container type
within constraints.

The first change now says that if the user does not supply a value for a
contrainer constraint type the model service will now set a default one
of None.

If a caller now attempts to set a container constraint that is not valid
an error indicating this will be returned.