chore: replace CharmTag usage in client/annotations facades
Since the removal of CharmTag in juju/names, we need to clean up its
only usage in juju's codebase.
This was done by simply checking the prefix of the entity tag and
treating it as a charm URL afterwards.
This behavior will change in a follow-up PR that will pass charm
locators to charm domain service methods.