fix(ci): fix relation departing unit test on aws
This patch fixes the relation departing unit CI test that was failing
only on AWS.
The reason is that after the patch
https://github.com/juju/juju/pull/18614 the logs are no longer received
for the removed unit, making the check on logs impossible for the test
to succeed.
The fix was therefore to check the values of the `related-units` field
in the yaml output of juju show-unit.