feat: ensure the charm does not change
We know that the charm config is immutable, so it's safe to use
the config, but there is a race which could happen between reading
the config and setting the charm config. In this situation, just
use the charm UUID when writing the transaction, which will guard
against any races.