chore: update charm domain methods to use natural keys
This patch replaces all the charm.ID input arguments in the charm
service methods with a charm.CharmLocator struct, which acts as a
natural key.

The charms should be identifiable with their name, source and revision.
Although CharmLocator also contains an Architecture field, we leave it
empty when passing the locator for the service methods.