fix(race): with unit assignment
During the transition from Mongo to DQLite, there is the possibility of
race conditions resulting from state being reading in one source, but
not ready in another.

One occurrence of this is which I have observed is during unit
assignment. The unit_assignment doc is placed into Mongo, before the
charm is ready in DQLite. This results in a "charm not found" error when
the unitassigner worker tries to place the unit.

Resolve this forcibly by placing a retry loop in state code (NOTE: This
part of state code is driven by our domain, we inject the charm service
into state as part of the transition). This code is dead code anyway,
that will be removed very soon, so a brute force solution like this is
not a concern.

These kinds of race conditions will resolve themselves automatically
when the transition from one state to another is completed.