feat: expose application config and settings
The following exposes the application config along with the settings
independently of the config. We'll place the settings into the
config for the application service, but keep the settings separated
for migration.

I've made most of the atomic methods deprecated. On going work to
remove this is being done.

The rename of ApplicationConfig to ApplicationConfigAndSettings
reflect the work involved.