fix(exec): for charms with no actions.yaml
Recently the core/charm package was refactored to reduce redundancy.
This included providing functions to read specific files from a charm
dir, meaning we don't need to parse the entire charm for a single bit of
data.

When one of our consumers, the uniter runner, which is used to run
actions, was refactored a bug was introduced. If we attempt to read the
actions for a charm which doesn't have one, the charm package raises a
FileNotFound error, which we forgot to handle in the uniter.

Handle this error by ignoring it