refactor: disable dashboardconnnectioninfo method
Rather that convert this method to use new domains while replacing
mongodb with dqlite, disable this method by returning NotImplemented.

The functionality needs to be moved to the controller charm as it is.
The charm's purpose is to expose configurations and integration
endpoints to work with the dashboard, haproxy and other charms.

CI dashboard test updated to check for the appropriate not implemented
message.