refactor: adds model constraint not found
Introduces a new model constraint error for not found where no model
constraints have been set for a model. This is a situation that arises
when a user hasn't previously set any constraints in the database.

Instead of making the state layer determine what to do in this case we
now have the state layer just tell us that no value is present and the
service layer can make the determination on what needs to be done.

This allows us to put business logic in the service and not have the
state layer understands higher level caller concerns.