refactor(ports): to remove RunAtomic calls
RunAtomic is an anti-pattern. We recently decided to remove it.

One of the places RunAtomic was used was in the unit ports domain. We
had the admirable aim of isolating the business logic, specifically how
juju interacts with the wildcard endpoint, away from persistence logic.

So removing RunAtomic meant moving all this logic out of the service
layer and into the state layer. UpdateUnitPorts is now very complex.

Move all the logic surrounding UpdateUnitPorts into it's own file. Do
the same for tests.

This was pretty much a 1-to-1 migration. However, there was one marginal
change. I discovered a bug with GetEndpointOpenedPorts, a method
designed to get the opened ports for a specific endpoint. The wildcard
is a special case endpoint, which broke this method. But it turned out,
now the method is private, we only need this method to query the
wildcard endpoint.

Rewire this method slightly to be specialised for the wildcard endpoint.