fix(resources): add back DeployRepository upload resources param
DeployFromRepository used to return a param that listed the added
pending resources to upload. This was removed when depreicating pending
resources. The problem is, this is also how the client works out how to
upload its local resources. To upload them it needs to know their type,
which was returned in this parameter.

Add the return value back in (was nil) to fix client.