feat(resource): support potential resources with undefined state
This commit introduces the concept of "potential" resources, which are resources belonging to an
application but not linkable to any units. Those resources are populated on the creation of
the application, as long as the charm is a repository charm, and are bound to be updated through
SetRepositoryResources function.

This function update the last_polled state of each potential resources it get as argument, and set
a new revision. It is meant to be used by revision worker in order to refresh potential updates for
a repository charm.