refactor: introduce views for constraints
Introduces new view for getting fully formed constraints and also the
constraints set on the model. This allows us to have cleaner and smaller
sql in the go files and be more explicit with what we want.

I have also started naming the view with an s for constraints instead of
constraint. This is part of a wider change to get the naming correct
across the board here.