chore: always pass architecture into charm locator from url on facades
This patch makes sure that every time we call a charm domain method with
a charm locator, the architecture is set from the values of the provided
charm url in the facade endpoint. It's not an issue if the charm url
does not contain an architecture, it will be empty when querying the db.