refactor(application): remove RunAtomic from CreateApplication
The application domain was constructed making liberal use of RunAtomic.
But we have since decided that this was an anti-pattern, so it should be
removed.

Re-write a few methods to not be atomic. Starting with
CreateApplication, but spilling over into a few others.

This was not a trivial lift-and-switch for in a few ways. Particularly
of note:
1. CreateApplication now has a multi-variadic parameter of
   AddUnitParams. This allows us to create an application and add units
   within the same transaction. A large number of instances of RunAtomic
   were doing exactly that.
2. InsertCAASUnit was created. The service method RegisterCAASUnit used
   to call InsertUnit in it's RunAtomic body. However, this body was
   very complex. So I have added a single method, InsertCAASUnit, on the
   state layer which now contains all this complexity.

As a flyby, I have also converted over a large number of errors.Annotate
to errors.Errorf