fix(charm): make `ReserveCharmRevision` reentrant
This commit allows to call `ReserveCharmRevision` twice on the same charm.
The first one would create a new charm and return the ID. The second one will
be a noop o the state, but still return the charm.ID.

It is necessary to allows to perform further updates on revision which rely on charm
identity, like linked resources.