feat: introduce model constraints importing
This commit introduces a new import operation for consuming and
importing the constraints set on a model's description during import. To
achieved this we have re-used the logic we had from the mongo
implementation around testing if a value has been set or not.

As part of this commit I have also moved the domain/model/modelmigration
package over to internal/errors.