fix: fix data race in workers
There is a potential for data races around the context in the
api remote caller. Getting the context cancellation right is a bit
tricky because of how we want the semantics to be for changes.

The solution is simple, put the cancellation in the goroutine, so
there is no possibility of a race. Everything is sequenced.