test: fix race with mocks in resource opener
The helper function to create a new unit resource opener expects mocked
methods. The helper function was called in concurrent goroutines created
in the test. This mean the mocks were being read and set at the same
time, which was a race.

This is fixed by seperating out the expect functions from the call to
the new opener.