refactor: allocate public ip address meaning
Because of the way the dqlite driver handles null bools we have to test
if the bool is also false before setting the value on constraints.
Otherwise we risk reporting to the user that a value is set when it
wasn't. This is a known problem with Dqlite and null bools being set.