chore: remove equinix metal provider
Equinix is sunsetting the metal offering[1], and what is more the
client SDK we interact with the provider is also made read-only[2].
This is a potential security concern that an unmaintained package
could have vulnerabilities in the future. Rewriting the provider
to use a new SDK seems like a long way off[3], even if they
weren't already sunsetting it.

  1. https://deploy.equinix.com/blog/sunsetting-equinix-metal/
  2. https://github.com/equinixmetal-archive/packngo
  3. https://github.com/equinix/equinix-sdk-go