feat(docs): autogen hook-command docs
We already auto-generate our reference docs for the juju cli using the
juju documentation command.

Do the same for the jujuc hook commands. Indirectly use the juju/cmd
documentation code.

This isn't ideal, as importing this code from a 3rd party library
doesn't give us much control. However, this is the easiest and most
consistant way to generate these docs.

Write a new function in conf.py similar to the function to generate the
cli docs & re-configure the directory to correctly display these docs in
sphinx.