refactor: remove status data tables
We don't need to model the status data tables. The data with in
the rows has no RI possible, they can be nested maps of maps, or
maps of slices, so we should just create JSON blobs and be done
remove the burden from us having to second guess that information.

We might want to impose a limit on the size of that information,
but for now do what we've always done.