feat: implement state layer for application constraints
This patch implements the state layer in the application domain for
application constraints.
Only two methods are needed:
- GetApplicationConstraints
- SetApplicationConstraints
It must be noted that in the case of referential integrity errors (i.e.
foreign key constraint errors), the method will return a bespoke error.
This was not the case juju worked up until 4.0, instead the constraints
were accepted but the machine was never started and errors were logged.