refactor(refresh): improve local refresh name checks
Juju does not allow refreshing charms to change the actual charm itself.

This is easy to enforce for charmhub, since we only allow refreshing to
change the channel, revision, etc. of the same charm in charmhub.

However, for local charm, which are user-controlled opaque-ish blobs to
the CLI, this is not possible to do so with rigour. A user could put
anything they like in a blob, so it's up to best guess.

One of the safety guard we put in place is we check the name of the
charm has not changed. We do this by querying the controller for the
charm url, and comparing that to the name of our local charm in
metadata.

However, charmhub charms are not necessarily named the same in charmhub
as they are in their metadata. For instance, juju-qa-dummy-subordinate
has that name in charmhub, but the name in metadata is just
dummy-subordinate. That means we cannot refresh the charm, even to an
identical local copy of itself.

This could be improved by querying the controller instead for the name
in metadata, not the charm url, but that is simply not worth it, since
we'd need a new facade endpoint for this very niche scenerio.

Instead, I have simply improved the warning, and allow you to skip the
check by using the pre-existing --force flag.