feat(relations): wire up dummy relations service to application facade
Wire up the NotImplemented relations service to the client application
facade.

The AddRelation facade is changed from having a seperate InferEndpoints
and AddRelation call to having an AddRelation and an AddPeerRelation
call.

Previously, the AddRelation and InferEndpoints calls had branching logic
depending on the number of variadic arguments, if there was one then it
was a peer relation, two then it was a normal relation.

In this change, a new method is added, AddPeerRelation. This allows the
branching on the number of arguments to be done once, at the application
facade level and makes the methods in the service cleaner.

The InferEndpoints method is not replaced in the domain, instead, the
logic it contains will be moved to the service layer of AddRelation and
AddPeerRelation.

The body of the DestroyRelation facade method is replaced with
NotImplemented, as this work will be done at a later data.