refactor: change event type
You never want a change type of created alone. Instead, you want
it to be both created and updated. This is because of coalescing
of events, the type has precedence: delete will always override
update, and update will alway override create. In typical workflows,
a created followed by an update is a real possibility, so if a
watcher expects to only see a created event, this is doomed to
failure. Instead you will **only** see an update. The fix here
is to hide created from the watchers and only expose Changed and
Deleted (along with All). By removing Create, we remove the chance
of falling into the trap of only seeing the created event.

Some of the watchers have had to be changed, as they depend on the
old concept, which means they aren't correct in their design and
have a potential bug waiting to happen.