chore: use INT instead of nullable bool for allocate_public_ip constraint
There is a bug on dqlite (or go-dqlite) that prevents us from using
(nullable) BOOLs since they always return as false. The workaround here
is to use an INT and set the table as STRICT, this way we can still use
a `bool` on our go struct and the driver will correctly continue mapping
it.
Add some more tests to make sure this behavior is OK.