feat(sshserverworker): revert graceful closure & pr comments
Adds a dirtykill check and cleankill check to the worker. We've revrted back to a forceful kill of
the worker as testing this is extremely difficult with a in memory listener. The in memory listener
shares the same connection object and closes both the dialer  connection and server listener
connection.  Should we require a need for graceful closures, we can investigate this in the future.