feat: wire up credential invalidator
This is the first steps towards removing the provider call context.
The provider already knows about this information, services shouldn't
be required to chase this up.

The solution here is to wire up the credential invalidator into the
providers when we open the environ (we can do this now because of the
provider tracker), and then pass in the invalidator once we're done.

I've not completely removed the ProviderCallContext, that can come
after this PR has landed. It should be a simple job of just swapping
it for a context.Context and then passing the invalidator directly
into the "MaybeHandleCredentialError" function.

I'm not sure of the need to centralise all of this yet, it seems
overkill, but I don't want to change the providers too much.