feat: implement application constraint service methods
This patch adds two domain service methods for retrieving and setting
application constraints.
Retrieval, which is done via `GetApplicationConstraints()` is very
simple and acts as a proxy for the state layer method.
Constraint setting on the other hand, which is done via
`SetApplicationConstraints()` performs the actual constraint validation
via the providers, as well as acting as a proxy (thus returning the
bespoke errors) to the state layer method.

Note that the tests added were inspired by the state/ application
constraint tests that were present in pre-4.0 Juju versions.