feat: introduce environs credential invalidator
This is the precursor to removing the ProviderCallContext, by adding
an invalidator to every provider, each provider will be allowed to
invalidate a credential if they hit an auth error.

This isn't added to the OpenParams for a Provider because we want
to ensure that we've not missed anything when compiling. We can
move this into the OpenParams struct once everything has been
migrated.

The next steps would be the following:

 1. Allow the ProviderTracker to hand out Providers without the
    ability to invalidate credentials.
 2. Move all providers to use the invalidator instead of the
    ProviderCallContext.
 3. Migrate workers to use the CredentialAPI as the invalidator
    when constructing the environ.
 4. Remove ProviderCallContext as it will be deadcode.

The code is relatively simple, it's intruduction of a new interface,
nothing calls it yet. It just sets out the plan of action.