docs: only match major.minor
Handle 2 things
a) local builds don't look quite like snap builds (3.6.3-ubuntu-amd64 vs
3.6.2-genericlinux-amd64)
b) allow for minor version but don't worry about patch

The earlier check was trying to make it harder to make a mistake and
build with a slightly wrong version, but it didn't really cover for the
fact that if you actually edited the source tree, but didn't compile a
new version, it wouldn't generate the right docs anyway. And it meant
that people that just wanted to touch the current tip of a branch would
fail to build the docs using a snap version of juju.